Rrequirements.txt
absl-py==0.6.1
aiml==0.9.2
astor==0.7.1
blinker==1.4
boto==2.49.0
boto3==1.9.35
botocore==1.12.35
bz2file==0.98
certifi==2018.10.15
chardet==3.0.4
ChatterBot==0.8.7
chatterbot-corpus==1.1.4
cycler==0.10.0
cymem==2.0.2
dateparser==0.7.0
dateutils==0.6.6
dill==0.2.8.2
Django==1.11
django-redis==4.10.0
django-redis-cache==1.8.1
django-rest-framework-mongoengine==3.3.1
django-session-timeout==0.0.3
djangorestframework==3.9.0
docutils==0.14
future==0.16.0
gast==0.2.0
gensim==3.6.0
grpcio==1.16.1
h5py==2.8.0
idna==2.7
jmespath==0.9.3
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
many-stop-words==0.2.2
Markdown==3.0.1
mathparse==0.1.2
matplotlib==3.0.1
mongoengine==0.15.3
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
nltk==3.3
numpy==1.15.3
oauthlib==2.1.0
Pillow==5.3.0
plac==0.9.6
preshed==2.0.1
protobuf==3.6.1
Pygments==2.2.0
pymongo==3.7.2
pyparsing==2.3.0
pyreadline==2.1
python-dateutil==2.6.1
python-twitter==3.4.2
pytz==2018.5
PyYAML==3.13
redis==2.10.6
regex==2018.1.10
requests==2.20.0
requests-oauthlib==1.0.0
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
smart-open==1.7.1
SQLAlchemy==1.2.12
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
tflearn==0.3.2
toolz==0.9.0
tzlocal==1.5.1
urllib3==1.24
Werkzeug==0.14.1
